backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
qrcode
Pillow
mysqlclient
django-cors-headers